% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_power.R
\name{mod_power}
\alias{mod_power}
\title{Compute Power for Categorical Moderation Meta-analysis}
\usage{
mod_power(
  n_groups,
  effect_sizes,
  sample_size,
  k,
  model = c("fixed", "random"),
  hg = c("small", "medium", "large"),
  p = 0.05,
  es_type = c("Correlation", "d", "OR"),
  test_type = c("two-tailed", "one-tailed"),
  sd_within
)
}
\arguments{
\item{n_groups}{Number of anticipated groups in moderation analysis.}

\item{effect_sizes}{Expected effect sizes of for each group.}

\item{sample_size}{Expected number of participants per group.}

\item{k}{Total expected number of studies.}

\item{model}{Fixed-effects model (model = "fixed") or Random-effects model (model = "random")}

\item{hg}{Expected heterogenity estimate (small, mdeium, large)}

\item{p}{alpha level: p = .05 (DEFAULT)}

\item{es_type}{Effect size metric: 'Correlation', 'd', or 'OR'.}

\item{test_type}{one-tailed or two-tailed}

\item{sd_within}{standard deviation within groups}
}
\value{
Estimated Power estimates for between and within-groups moderation
}
\description{
mod_power( ) is an extension of mpower() and so it takes similiar inputs. The primary inputs are the effected effect sizes
for each group and the expected within-group standard deviation in those groups
}
\examples{
mod_power(3, c(0,.1,.55), sample_size = 15, k = 15,
model = "random", hg = "small",p = .05, es_type = "Correlation",
sd_within = c(1,1,4), test_type = "two-tailed")

}
