% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpower.R
\name{mpower}
\alias{mpower}
\title{Compute Power for Meta-analysis}
\usage{
mpower(effect_size, sample_size, k, hg, model, test_type, p = 0.05)
}
\arguments{
\item{effect_size}{anticipated magnitude of effect size.}

\item{sample_size}{anticipated average number of participants per group}

\item{k}{anticipated number of studies}

\item{hg}{anticipated heterogenity estimate (smalll = 1.33, moderate = 1.67, large = 2)}

\item{model}{Fixed-effects model (model = "fixed") or Random-effects model (model = "random")}

\item{test_type}{one-tailed or two-tailed}

\item{p}{alpha level: p = .05 (DEFAULT)}
}
\value{
Estimated power
}
\description{
the mpower() command allows for the calculation of statistical power for both fixed- and random-effects
meta-analyis models. Currently, the anticpated summary effect size estimate should be standardized mean differences units
(i.e, Cohen's d)
}
