% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_homogen_mpower.R
\name{plot_homogen_mpower}
\alias{plot_homogen_mpower}
\title{Plot Power Curve for Test of Homogeneity}
\usage{
plot_homogen_mpower(obj)
}
\arguments{
\item{obj}{should be an mpower object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plot Power Curve for Test of Homogeneity
}
