% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_homogen_power.R
\name{plot_homogen_power}
\alias{plot_homogen_power}
\title{Plot Power Curve for Test of Homogeneity}
\usage{
plot_homogen_power(obj)
}
\arguments{
\item{obj}{should be an homogen_power object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plot Power Curve for Test of Homogeneity
}
