% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod_power.R
\name{plot_mod_power}
\alias{plot_mod_power}
\title{Plot Power Curve for Moderator analysis}
\usage{
plot_mod_power(obj)
}
\arguments{
\item{obj}{This should be an mpower object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plot Power Curve for Moderator analysis
}
