% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modpower.R
\name{plot_modpower}
\alias{plot_modpower}
\title{Plot Power Curve for Categorical Moderator Analysis}
\usage{
plot_modpower(modpower)
}
\arguments{
\item{modpower}{This should be an "modpower" object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plot Power Curve for Categorical Moderator Analysis
}
