% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mpower.R
\name{plot_mpower}
\alias{plot_mpower}
\title{Plot Power Curve for Meta-analysis}
\usage{
plot_mpower(obj)
}
\arguments{
\item{obj}{This should be an "mpower" object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plots power curves for fixed effects models with various effect size magnitudes. Also plots power curves for various levels
of heterogeneity (e.g., i2 = 75%)
}
