% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{power_plot}
\alias{power_plot}
\title{Plot Power Curve for Meta-analysis}
\usage{
power_plot(obj)
}
\arguments{
\item{obj}{This should be an mpower object}
}
\value{
Power curve plot for the user specified input parameters
}
\description{
Plot Power Curve for Meta-analysis
}
