% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_metapower.R
\name{shiny_metapower}
\alias{shiny_metapower}
\title{Launch shiny application of metapower}
\usage{
shiny_metapower()
}
\value{
Launches shiny application
}
\description{
Calling this functioning simply launches the shiny application for metapower,
which allows users to easily compute power analysis with a shiny application
}
\examples{
shiny_metapower()


}
