% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_power.R
\name{mod_power}
\alias{mod_power}
\title{Compute Power for Categorical Moderation Meta-analysis}
\usage{
mod_power(
  n_groups,
  effect_sizes,
  sample_size,
  k,
  es_type,
  model,
  hg,
  test_type,
  p,
  sd_within,
  con_table
)
}
\arguments{
\item{n_groups}{Number of anticipated groups in moderation analysis.}

\item{effect_sizes}{Expected effect sizes of for each group.}

\item{sample_size}{Expected number of participants per group.}

\item{k}{Total expected number of studies.}

\item{es_type}{Effect size metric: 'Correlation', 'd', or 'OR'.}

\item{model}{Fixed-effects model (model = "fixed") or Random-effects model (model = "random")}

\item{hg}{Expected heterogenity estimate (small, mdeium, large)}

\item{test_type}{one-tailed or two-tailed}

\item{p}{alpha level: p = .05 (DEFAULT)}

\item{sd_within}{standard deviation within groups}

\item{con_table}{Only relevant for Odds Ratio. Expected 2x2 contingency table as a vector in the following format: c(a,b,c,d)

\tabular{lcc}{
 2x2 Table   \tab Group 1 \tab Group 2 \cr
 Present     \tab a       \tab b       \cr
 Not Present \tab c       \tab d       \cr
}}
}
\value{
Estimated Power estimates for between and within-groups moderation
}
\description{
mod_power( ) is an extension of mpower() and so it takes similiar inputs. The primary inputs are the effected effect sizes
for each group and the expected within-group standard deviation in those groups
}
\examples{
mod_power(
 n_groups = 3,
 effect_sizes = c(0,.1,.55),
 sample_size = 15,
 k = 15,
 model = "random",
 hg = "small",
 es_type = "Correlation",
 sd_within = c(1,1,4),
 test_type = "two-tailed",
 p = .05)

}
