% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_power.R
\name{mod_power}
\alias{mod_power}
\title{Compute Power for Categorical Moderator Analysis in Meta-analysis}
\usage{
mod_power(
  n_groups,
  effect_sizes,
  study_size,
  k,
  i2,
  es_type,
  p = 0.05,
  con_table = NULL
)
}
\arguments{
\item{n_groups}{Numerical value for the levels of a categorical variable.}

\item{effect_sizes}{Numerical values for effect sizes of for each group.}

\item{study_size}{Numerical value for number of participants (per study).}

\item{k}{Numerical value for total number of studies.}

\item{i2}{Numerical value for Heterogeneity estimate (i^2).}

\item{es_type}{Character reflecting effect size metric: 'r', 'd', or 'or'.}

\item{p}{Numerical value for significance level (Type I error probability).}

\item{con_table}{(Optional) List of numerical values for 2x2 contingency
    tables as a vector in the following format: c(a,b,c,d). These should be
    specified for each group(i.e., n_groups).

\tabular{lcc}{
 2x2 Table   \tab Group 1 \tab Group 2 \cr
 Present     \tab a       \tab b       \cr
 Not Present \tab c       \tab d       \cr
}}
}
\value{
Estimated Power estimates for moderator analysis under fixed- and random-effects models
}
\description{
Computes statistical power for categorical moderator analysis under fixed and random effects models.
}
\examples{
mod_power(n_groups = 2,
          effect_sizes = c(.1,.5),
          study_size = 20,
          k = 10,
          i2 = .50,
          es_type = "d")
mod_power(n_groups = 2,
          con_table = list(g1 = c(6,5,4,5), g2 = c(8,5,2,5)),
          study_size = 40,
          k = 20,
          i2 = .50,
          es_type = "or")

}
\seealso{
\url{https://jason-griffin.shinyapps.io/shiny_metapower/}
}
