% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpower.R
\name{mpower}
\alias{mpower}
\title{Compute Power for Meta-analysis}
\usage{
mpower(
  effect_size,
  sample_size,
  k,
  es_type,
  model,
  hg,
  test_type = "two-tailed",
  p = 0.05,
  sd
)
}
\arguments{
\item{effect_size}{anticipated magnitude of effect size.}

\item{sample_size}{anticipated average number of participants per group}

\item{k}{anticipated number of studies}

\item{es_type}{Effect size metric: 'Correlation', 'd', or 'OR'}

\item{model}{Fixed-effects model (model = "fixed") or Random-effects model (model = "random")}

\item{hg}{anticipated heterogenity estimate (smalll = 1.33, moderate = 1.67, large = 2)}

\item{test_type}{one-tailed or two-tailed}

\item{p}{alpha level: p = .05 (DEFAULT)}

\item{sd}{expected difference in effect sizes}
}
\value{
Estimated Power
}
\description{
the mpower( ) command allows for the calculation of statistical power for both fixed- and random-effects
meta-analyis models. Currently, the anticpated summary effect size estimate should be standardized mean differences units
(i.e, Cohen's d)
}
\examples{
mpower(effect_size = .5, sample_size = 10, k = 10, hg = "large", es_type = "d",
       model = "random", test_type = "two-tailed", sd = .5)

}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P. T. and Rothstein, H. R.(2009). Introduction to meta-analysis, Chichester, UK: Wiley.

Hedges, L., Pigott, T. (2004). The Power of Statistical Tests for Moderators in Meta-Analysis Psychological Methods  9(4), 426-445.
doi: https://dx.doi.org/10.1037/1082-989x.9.4.426

Pigott, T. (2012). Advances in Meta-Analysis.
doi: https://dx.doi.org/10.1007/978-1-4614-2278-5
}
