% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpower.R
\name{mpower}
\alias{mpower}
\title{Compute Power for Meta-analysis}
\usage{
mpower(
  effect_size,
  sample_size,
  k,
  es_type,
  model,
  hg,
  test_type = "two-tailed",
  p = 0.05,
  sd,
  con_table
)
}
\arguments{
\item{effect_size}{Expected effect size magnitude.}

\item{sample_size}{Expected number of participants (per group).}

\item{k}{Expected number of studies.}

\item{es_type}{Effect size metric: 'Correlation', 'd', or 'OR'}

\item{model}{Fixed-effects model (model = "fixed") or Random-effects model (model = "random")}

\item{hg}{Expected heterogenity: "small", "medium", or "large".}

\item{test_type}{Alternative hypothesis: "two-tailed" (default) or "one-tailed"}

\item{p}{alpha level: p = .05 (default)}

\item{sd}{(Optional) Fixed-effects models only: Expected standard deviation of among all effect sizes}

\item{con_table}{Only relevant for Odds Ratio. Expected 2x2 contingency table as a vector in the following format: c(a,b,c,d)

\tabular{lcc}{
 2x2 Table   \tab Group 1 \tab Group 2 \cr
 Present     \tab a       \tab b       \cr
 Not Present \tab c       \tab d       \cr
}}
}
\value{
Estimated Power
}
\description{
the mpower( ) command allows for the calculation of statistical power for both fixed- and random-effects
meta-analyis models. Currently, the anticpated summary effect size estimate should be standardized mean differences units
(i.e, Cohen's d)
}
\examples{
mpower(effect_size = .5, sample_size = 10, k = 10, es_type = "d",
       hg = "large", model = "random", test_type = "two-tailed")

}
\references{
Bornstein, M., Hedges, L. V., Higgins, J. P. T. and Rothstein, H. R.(2009). Introduction to meta-analysis, Chichester, UK: Wiley.

Hedges, L., Pigott, T. (2004). The Power of Statistical Tests for Moderators in Meta-Analysis Psychological Methods  9(4), 426-445.
doi: https://dx.doi.org/10.1037/1082-989x.9.4.426

Pigott, T. (2012). Advances in Meta-Analysis.
doi: https://dx.doi.org/10.1007/978-1-4614-2278-5
}
