% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mod_power.R
\name{plot_mod_power}
\alias{plot_mod_power}
\title{Plot Power Curve for Categorical Moderators}
\usage{
plot_mod_power(obj)
}
\arguments{
\item{obj}{This should be an 'mod_power' object}
}
\value{
Power curves for moderator analysis under fixed and random effects models
}
\description{
Plots power curves for categorical moderator in meta-analysis
}
