% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subgroup_power.R
\name{plot_subgroup_power}
\alias{plot_subgroup_power}
\title{Plot Power Curve for Subgroup analysis}
\usage{
plot_subgroup_power(obj)
}
\arguments{
\item{obj}{This should be an 'subgroup_power' object}
}
\value{
Power curves to detect subgroup differences for fixed and random effects models
}
\description{
Plots power curves to detect subgroup differences in meta-analysis.
}
